refactor(api): carry guard config on the oRPC context; move ports to orpc/services
`configureORPC()` was the last module-level registry: the guards read the
rate-limit budget, project id and AI key material from a process singleton the
host filled at boot. They now read `context.config`, which is required on
`BaseOSContext` — `apps/service` builds it from env once in `orpc.factory.ts`,
the dashboard's in-process client names its own budget. `orpc/config.ts` is
deleted.

The two host port modules move under `packages/api/orpc/services/`, mirroring
`routes/` and `guards/`: `agent-service.ts` → `services/agent.service.ts`,
`indexing.ts` → `services/indexing.service.ts`.
