+++ AGENTS.md
@@ -48,7 +48,7 @@ It doubles as the place where new stack and architecture ideas get tried for rea
 
 **Guards and policies.** Authorization logic lives once in `packages/service-kit/src/policies` (`sessionPolicy`, `apiKeyPolicy`, `adminPolicy`, `rateLimitPolicy`, `captchaPolicy`, `aiKeyPolicy`) and is bound to each transport by a thin adapter: `toHonoMiddleware` for Hono middleware (`apps/service/src/guards/`), `runPolicy` for oRPC middleware (`packages/api/orpc/guards/`). Write new authorization as a policy, not as a guard.
 
-**Config injection.** `packages/api` parses no env of its own. The hosting app calls `configureORPC({ rateLimit, projectId, aiAuthPrivateKey })` once at boot (`apps/service/src/factories/orpc.factory.ts`). Domain ports travel on the oRPC context (`BaseOSContext` in `packages/api/orpc/utils.ts`): `hooks.onFeedChanged` / `onFeedRemoved`, `indexing`, `agentKinds`. `createORPCContext` in `orpc.factory.ts` is the one place `apps/service` supplies them; a process that leaves one out (the dashboard's in-process router client) gets `SERVICE_UNAVAILABLE` from the routes that need it. Anything needing a long-lived process, a DB handle, or gateway credentials belongs in the app, not in `packages/api`.
+**Context injection.** `packages/api` parses no env of its own. Everything the guards and routes need from the host travels on the oRPC context (`BaseOSContext` in `packages/api/orpc/utils.ts`): `config` (rate-limit budget, project id, AI key material — required), plus the optional domain ports `hooks.onFeedChanged` / `onFeedRemoved`, `indexing` and `agentKinds`. The port interfaces live in `packages/api/orpc/services/`. `createORPCContext` in `apps/service/src/factories/orpc.factory.ts` is the one place `apps/service` supplies them; the dashboard's in-process router client supplies `config` and leaves the ports out, so the routes that need one answer `SERVICE_UNAVAILABLE`. Anything needing a long-lived process, a DB handle, or gateway credentials belongs in the app, not in `packages/api`.
 
 **Data access.** oRPC handlers never write raw Drizzle queries; they call repositories exported as `@chia/db/repos/*`. Write logic shared with workflow steps lives in `packages/api/<domain>/write` so a durable turn can call it without a request to authorize against.
